+++ src/TeaPages.php
@@ -17,7 +17,7 @@ if (!defined('TTO_CONTEXT')) {
  * @package Tea Theme Options
  * @subpackage Tea Pages
  * @author Achraf Chouk <ach@takeatea.com>
- * @since 2.3.0
+ * @since 2.3.3
  *
  */
 class TeaPages
@@ -99,7 +99,7 @@ class TeaPages
      * @param string $identifier Define the main slug
      * @param array $options Define if we can display connections and elasticsearch pages
      *
-     * @since 1.5.2.8
+     * @since 2.3.3
      */
     public function __construct($identifier, $options)
     {
@@ -116,7 +116,7 @@ class TeaPages
             );
         }
 
-        // Define parameters
+        //Define parameters
         $this->identifier = $identifier;
         $this->can_upload = current_user_can('upload_files');
 
@@ -156,6 +156,10 @@ class TeaPages
             //...Or update CPTs options...
             elseif ('cpts' == $for) {
                 $this->updateCpts();
+            }
+            //...Or update rewrite rules...
+            elseif ('rewrites' == $for) {
+                $this->updateRewrites();
             }
             //...Or update options...
             elseif ('settings' == $for) {
@@ -212,7 +216,7 @@ class TeaPages
     /**
      * Hook building scripts.
      *
-     * @uses wp_enqueue_media()
+     * @uses wp_enqueue_media_tto()
      * @uses wp_enqueue_script()
      *
      * @since 1.5.0-1
@@ -227,15 +231,15 @@ class TeaPages
         $jq = array('jquery');
 
         //Enqueue media and colorpicker scripts
-        if (function_exists('wp_enqueue_media')) {
-            wp_enqueue_media();
+        //if (function_exists('wp_enqueue_media_tto')) {
+            $this->wp_enqueue_media_tto();
             wp_enqueue_script('wp-color-picker');
             wp_enqueue_script('accordion');
-        }
-        else {
-            wp_enqueue_script('media-upload');
-            wp_enqueue_script('farbtastic');
-        }
+        //}
+        //else {
+            //wp_enqueue_script('media-upload');
+            //wp_enqueue_script('farbtastic');
+        //}
 
         //Enqueue all minified scripts
         wp_enqueue_script('tea-to', TTO_URI.'/assets/js/teato.min.js', $jq);
@@ -723,7 +727,7 @@ class TeaPages
     /**
      * Build footer layout.
      *
-     * @since 1.5.2.1
+     * @since 2.3.3
      */
     protected function buildLayoutFooter()
     {
@@ -740,6 +744,7 @@ class TeaPages
             ? admin_url('admin.php?page='.$this->identifier.'&action=tea_action&for=caps')
             : '';
         $cpturl = admin_url('admin.php?page='.$this->identifier.'&action=tea_action&for=cpts');
+        $rwturl = admin_url('admin.php?page='.$this->identifier.'&action=tea_action&for=rewrites');
 
         //Include template
         include(TTO_PATH.'/Tpl/layouts/__layout_footer.tpl.php');
@@ -1094,6 +1099,29 @@ class TeaPages
     }
 
     /**
+     * Update rewrite rules options.
+     *
+     * @param boolean $redirect Define if the TTO has to make a redirect
+     *
+     * @since 2.3.3
+     */
+    protected function updateRewrites($redirect = true)
+    {
+        //Check if we are in admin panel
+        if (!TTO_IS_ADMIN) {
+            return;
+        }
+
+        //Flush all rewrite rules
+        flush_rewrite_rules();
+
+        //Redirect to Tea TO homepage
+        if ($redirect) {
+            wp_safe_redirect(admin_url('admin.php?page='.$this->identifier.'&action=tea_action&for=dashboard'));
+        }
+    }
+
+    /**
      * @return Elasticsearch|void
      */
     protected function registerElasticSearch()
@@ -1105,5 +1133,324 @@ class TeaPages
         $field->setCurrentPage($page);
 
         return $field;
+    }
+
+
+
+
+
+
+
+
+
+
+
+
+    /**
+     * Update rewrite rules options.
+     *
+     * @param array $args Define if the TTO has to make a redirect
+     *
+     * @since 2.3.3
+     */
+    protected function wp_enqueue_media_tto($args = array())
+    {
+        //Check if we are in admin panel
+        if (!TTO_IS_ADMIN) {
+            return;
+        }
+
+        //Enqueue me just once per page, please.
+        if (did_action('wp_enqueue_media_tto')) {
+            return;
+        }
+
+        global $content_width, $wpdb, $wp_locale;
+
+        $defaults = array(
+            'post' => null,
+        );
+        $args = wp_parse_args($args, $defaults);
+
+        //We're going to pass the old thickbox media tabs to `media_upload_tabs`
+        //to ensure plugins will work. We will then unset those tabs.
+        $tabs = array(
+            //handler action suffix => tab label
+            'type'     => '',
+            'type_url' => '',
+            'gallery'  => '',
+            'library'  => '',
+        );
+
+        /** This filter is documented in wp-admin/includes/media.php */
+        $tabs = apply_filters('media_upload_tabs', $tabs);
+        unset($tabs['type'], $tabs['type_url'], $tabs['gallery'], $tabs['library']);
+
+        $props = array(
+            'link'  => get_option('image_default_link_type'), //db default is 'file'
+            'align' => get_option('image_default_align'), //empty default
+            'size'  => get_option('image_default_size'),  //empty default
+        );
+
+        $exts = array_merge(wp_get_audio_extensions(), wp_get_video_extensions());
+        $mimes = get_allowed_mime_types();
+        $ext_mimes = array();
+        foreach ($exts as $ext) {
+            foreach ($mimes as $ext_preg => $mime_match) {
+                if (preg_match('#' . $ext . '#i', $ext_preg)) {
+                    $ext_mimes[ $ext ] = $mime_match;
+                    break;
+                }
+            }
+        }
+
+        if ( false === ( $has_audio = get_transient( 'has_audio' ) ) ) {
+                $has_audio = (bool) $wpdb->get_var( "
+                        SELECT ID
+                        FROM $wpdb->posts
+                        WHERE post_type = 'attachment'
+                        AND post_mime_type LIKE 'audio%'
+                        LIMIT 1
+                " );
+                set_transient( 'has_audio', $has_audio );
+        }
+        if ( false === ( $has_video = get_transient( 'has_video' ) ) ) {
+                $has_video = (bool) $wpdb->get_var( "
+                        SELECT ID
+                        FROM $wpdb->posts
+                        WHERE post_type = 'attachment'
+                        AND post_mime_type LIKE 'video%'
+                        LIMIT 1
+                " );
+                set_transient( 'has_video', $has_video );
+        }
+        $months = $wpdb->get_results($wpdb->prepare("
+            SELECT DISTINCT YEAR(post_date) AS year, MONTH(post_date) AS month
+            FROM $wpdb->posts
+            WHERE post_type = %s
+            ORDER BY post_date DESC
+        ", 'attachment'));
+        foreach ($months as $month_year) {
+            $month_year->text = sprintf(__('%1$s %2$d'), $wp_locale->get_month($month_year->month), $month_year->year);
+        }
+
+        $settings = array(
+            'tabs'      => $tabs,
+            'tabUrl'    => add_query_arg(array('chromeless' => true), admin_url('media-upload.php')),
+            'mimeTypes' => wp_list_pluck(get_post_mime_types(), 0),
+            /** This filter is documented in wp-admin/includes/media.php */
+            'captions'  => ! apply_filters('disable_captions', ''),
+            'nonce'     => array(
+                'sendToEditor' => wp_create_nonce('media-send-to-editor'),
+           ),
+            'post'    => array(
+                'id' => 0,
+           ),
+            'defaultProps' => $props,
+            'attachmentCounts' => array(
+                'audio' => ($has_audio) ? 1 : 0,
+                'video' => ($has_video) ? 1 : 0
+           ),
+            'embedExts'    => $exts,
+            'embedMimes'   => $ext_mimes,
+            'contentWidth' => $content_width,
+            'months'       => $months,
+            'mediaTrash'   => MEDIA_TRASH ? 1 : 0
+        );
+
+        $post = null;
+        if (isset($args['post'])) {
+            $post = get_post($args['post']);
+            $settings['post'] = array(
+                'id' => $post->ID,
+                'nonce' => wp_create_nonce('update-post_' . $post->ID),
+           );
+
+            $thumbnail_support = current_theme_supports('post-thumbnails', $post->post_type) && post_type_supports($post->post_type, 'thumbnail');
+            if (! $thumbnail_support && 'attachment' === $post->post_type && $post->post_mime_type) {
+                if (wp_attachment_is('audio', $post)) {
+                    $thumbnail_support = post_type_supports('attachment:audio', 'thumbnail') || current_theme_supports('post-thumbnails', 'attachment:audio');
+                } elseif (wp_attachment_is('video', $post)) {
+                    $thumbnail_support = post_type_supports('attachment:video', 'thumbnail') || current_theme_supports('post-thumbnails', 'attachment:video');
+                }
+            }
+
+            if ($thumbnail_support) {
+                $featured_image_id = get_post_meta($post->ID, '_thumbnail_id', true);
+                $settings['post']['featuredImageId'] = $featured_image_id ? $featured_image_id : -1;
+            }
+        }
+
+        $hier = $post && is_post_type_hierarchical($post->post_type);
+
+        $strings = array(
+            //Generic
+            'url'         => __('URL'),
+            'addMedia'    => __('Add Media'),
+            'search'      => __('Search'),
+            'select'      => __('Select'),
+            'cancel'      => __('Cancel'),
+            'update'      => __('Update'),
+            'replace'     => __('Replace'),
+            'remove'      => __('Remove'),
+            'back'        => __('Back'),
+            /* translators: This is a would-be plural string used in the media manager.
+               If there is not a word you can use in your language to avoid issues with the
+               lack of plural support here, turn it into "selected: %d" then translate it.
+             */
+            'selected'    => __('%d selected'),
+            'dragInfo'    => __('Drag and drop to reorder media files.'),
+
+            //Upload
+            'uploadFilesTitle'  => __('Upload Files'),
+            'uploadImagesTitle' => __('Upload Images'),
+
+            //Library
+            'mediaLibraryTitle'      => __('Media Library'),
+            'insertMediaTitle'       => __('Insert Media'),
+            'createNewGallery'       => __('Create a new gallery'),
+            'createNewPlaylist'      => __('Create a new playlist'),
+            'createNewVideoPlaylist' => __('Create a new video playlist'),
+            'returnToLibrary'        => __('&#8592; Return to library'),
+            'allMediaItems'          => __('All media items'),
+            'allDates'               => __('All dates'),
+            'noItemsFound'           => __('No items found.'),
+            'insertIntoPost'         => $hier ? __('Insert into page') : __('Insert into post'),
+            'unattached'             => __('Unattached'),
+            'trash'                  => _x('Trash', 'noun'),
+            'uploadedToThisPost'     => $hier ? __('Uploaded to this page') : __('Uploaded to this post'),
+            'warnDelete'             => __("You are about to permanently delete this item.\n  'Cancel' to stop, 'OK' to delete."),
+            'warnBulkDelete'         => __("You are about to permanently delete these items.\n  'Cancel' to stop, 'OK' to delete."),
+            'warnBulkTrash'          => __("You are about to trash these items.\n  'Cancel' to stop, 'OK' to delete."),
+            'bulkSelect'             => __('Bulk Select'),
+            'cancelSelection'        => __('Cancel Selection'),
+            'trashSelected'          => __('Trash Selected'),
+            'untrashSelected'        => __('Untrash Selected'),
+            'deleteSelected'         => __('Delete Selected'),
+            'deletePermanently'      => __('Delete Permanently'),
+            'apply'                  => __('Apply'),
+            'filterByDate'           => __('Filter by date'),
+            'filterByType'           => __('Filter by type'),
+            'searchMediaLabel'       => __('Search Media'),
+            'noMedia'                => __('No media attachments found.'),
+
+            //Library Details
+            'attachmentDetails'  => __('Attachment Details'),
+
+            //From URL
+            'insertFromUrlTitle' => __('Insert from URL'),
+
+            //Featured Images
+            'setFeaturedImageTitle' => __('Set Featured Image'),
+            'setFeaturedImage'    => __('Set featured image'),
+
+            //Gallery
+            'createGalleryTitle' => __('Create Gallery'),
+            'editGalleryTitle'   => __('Edit Gallery'),
+            'cancelGalleryTitle' => __('&#8592; Cancel Gallery'),
+            'insertGallery'      => __('Insert gallery'),
+            'updateGallery'      => __('Update gallery'),
+            'addToGallery'       => __('Add to gallery'),
+            'addToGalleryTitle'  => __('Add to Gallery'),
+            'reverseOrder'       => __('Reverse order'),
+
+            //Edit Image
+            'imageDetailsTitle'     => __('Image Details'),
+            'imageReplaceTitle'     => __('Replace Image'),
+            'imageDetailsCancel'    => __('Cancel Edit'),
+            'editImage'             => __('Edit Image'),
+
+            //Crop Image
+            'chooseImage' => __('Choose Image'),
+            'selectAndCrop' => __('Select and Crop'),
+            'skipCropping' => __('Skip Cropping'),
+            'cropImage' => __('Crop Image'),
+            'cropYourImage' => __('Crop your image'),
+            'cropping' => __('Cropping&hellip;'),
+            'suggestedDimensions' => __('Suggested image dimensions:'),
+            'cropError' => __('There has been an error cropping your image.'),
+
+            //Edit Audio
+            'audioDetailsTitle'     => __('Audio Details'),
+            'audioReplaceTitle'     => __('Replace Audio'),
+            'audioAddSourceTitle'   => __('Add Audio Source'),
+            'audioDetailsCancel'    => __('Cancel Edit'),
+
+            //Edit Video
+            'videoDetailsTitle'     => __('Video Details'),
+            'videoReplaceTitle'     => __('Replace Video'),
+            'videoAddSourceTitle'   => __('Add Video Source'),
+            'videoDetailsCancel'    => __('Cancel Edit'),
+            'videoSelectPosterImageTitle' => __('Select Poster Image'),
+            'videoAddTrackTitle'    => __('Add Subtitles'),
+
+            //Playlist
+            'playlistDragInfo'    => __('Drag and drop to reorder tracks.'),
+            'createPlaylistTitle' => __('Create Audio Playlist'),
+            'editPlaylistTitle'   => __('Edit Audio Playlist'),
+            'cancelPlaylistTitle' => __('&#8592; Cancel Audio Playlist'),
+            'insertPlaylist'      => __('Insert audio playlist'),
+            'updatePlaylist'      => __('Update audio playlist'),
+            'addToPlaylist'       => __('Add to audio playlist'),
+            'addToPlaylistTitle'  => __('Add to Audio Playlist'),
+
+            //Video Playlist
+            'videoPlaylistDragInfo'    => __('Drag and drop to reorder videos.'),
+            'createVideoPlaylistTitle' => __('Create Video Playlist'),
+            'editVideoPlaylistTitle'   => __('Edit Video Playlist'),
+            'cancelVideoPlaylistTitle' => __('&#8592; Cancel Video Playlist'),
+            'insertVideoPlaylist'      => __('Insert video playlist'),
+            'updateVideoPlaylist'      => __('Update video playlist'),
+            'addToVideoPlaylist'       => __('Add to video playlist'),
+            'addToVideoPlaylistTitle'  => __('Add to Video Playlist'),
+        );
+
+        /**
+         * Filter the media view settings.
+         *
+         * @since 3.5.0
+         *
+         * @param array   $settings List of media view settings.
+         * @param WP_Post $post     Post object.
+         */
+        $settings = apply_filters('media_view_settings', $settings, $post);
+
+        /**
+         * Filter the media view strings.
+         *
+         * @since 3.5.0
+         *
+         * @param array   $strings List of media view strings.
+         * @param WP_Post $post    Post object.
+         */
+        $strings = apply_filters('media_view_strings', $strings,  $post);
+
+        $strings['settings'] = $settings;
+
+        //Ensure we enqueue media-editor first, that way media-views is
+        //registered internally before we try to localize it. see #24724.
+        wp_enqueue_script('media-editor');
+        wp_localize_script('media-views', '_wpMediaViewsL10n', $strings);
+
+        wp_enqueue_script('media-audiovideo');
+        wp_enqueue_style('media-views');
+        if (is_admin()) {
+            wp_enqueue_script('mce-view');
+            wp_enqueue_script('image-edit');
+        }
+        wp_enqueue_style('imgareaselect');
+        wp_plupload_default_settings();
+
+        require_once ABSPATH . WPINC . '/media-template.php';
+        add_action('admin_footer', 'wp_print_media_templates');
+        add_action('wp_footer', 'wp_print_media_templates');
+        add_action('customize_controls_print_footer_scripts', 'wp_print_media_templates');
+
+        /**
+         * Fires at the conclusion of wp_enqueue_media_tto().
+         *
+         * @since 3.5.0
+         */
+        do_action('wp_enqueue_media_tto');
     }
 }